refactor: rename modules for code validation to match spec
