+++ src/validation/instructions/mod.rs
@@ -1,2 +1,2 @@
-pub mod code;
-pub mod const_expr;
+pub mod constant_expressions;
+pub mod expressions;
