+++ src/validation/modules/globals.rs
@@ -9,7 +9,7 @@ use crate::{
         },
     },
     validation::{
-        instructions::const_expr::decode_and_validate_constant_expression,
+        instructions::constant_expressions::decode_and_validate_constant_expression,
         validation_stack::ValidationStack,
     },
     GlobalType, ValidationError,
