+++ src/validation/modules/mod.rs
@@ -1,6 +1,7 @@
 pub mod data_segments;
 pub mod element_segments;
 pub mod exports;
+pub mod functions;
 pub mod globals;
 pub mod imports;
 pub mod indices;
